research work
This repository accompanies our research paper on Explainable Artificial Intelligence (XAI), providing an in-depth survey and comparative analysis of modern explain-ability techniques in machine learning. In a time when machine learning models often operate as "black boxes", this work highlights the importance of transparency, interpretability, and accountability in AI systems.